+++ activerecord/test/cases/associations/eager_test.rb
@@ -826,18 +826,6 @@ class EagerAssociationTest < ActiveRecord::TestCase
     )
   end
 
-  def test_preload_with_interpolation
-    assert_deprecated do
-      post = Post.includes(:comments_with_interpolated_conditions).find(posts(:welcome).id)
-      assert_equal [comments(:greetings)], post.comments_with_interpolated_conditions
-    end
-
-    assert_deprecated do
-      post = Post.joins(:comments_with_interpolated_conditions).find(posts(:welcome).id)
-      assert_equal [comments(:greetings)], post.comments_with_interpolated_conditions
-    end
-  end
-
   def test_polymorphic_type_condition
     post = Post.all.merge!(:includes => :taggings).find(posts(:thinking).id)
     assert post.taggings.include?(taggings(:thinking_general))
@@ -1294,23 +1282,22 @@ class EagerAssociationTest < ActiveRecord::TestCase
     assert_equal pets(:parrot), Owner.including_last_pet.first.last_pet
   end
 
-  test "include instance dependent associations is deprecated" do
+  test "preloading and eager loading of instance dependent associations is not supported" do
     message = "association scope 'posts_with_signature' is"
-    assert_deprecated message do
-      begin
-        Author.includes(:posts_with_signature).to_a
-      rescue NoMethodError
-        # it's expected that preloading of this association fails
-      end
+    error = assert_raises(ArgumentError) do
+      Author.includes(:posts_with_signature).to_a
     end
+    assert_match message, error.message
 
-    assert_deprecated message do
-      Author.preload(:posts_with_signature).to_a rescue NoMethodError
+    error = assert_raises(ArgumentError) do
+      Author.preload(:posts_with_signature).to_a
     end
+    assert_match message, error.message
 
-    assert_deprecated message do
+    error = assert_raises(ArgumentError) do
       Author.eager_load(:posts_with_signature).to_a
     end
+    assert_match message, error.message
   end
 
   test "preloading readonly association" do